LayoutTests/imported/w3c/web-platform-tests/css/css-highlight-api/Highlight-setlike-tampered-Set-prototype-expected.txt


PASS Highlight is a setlike interface that works as expected even if Set.prototype is tampered.
